% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.moml.R
\name{et.moml}
\alias{et.moml}
\title{MOML with ET-Lasso}
\usage{
et.moml(
  x,
  treatment,
  reward,
  propensity_score,
  intercept = TRUE,
  weight = NULL,
  loss = c("logistic", "boost", "hinge-boost", "lum"),
  control = list(),
  nstages = 2,
  refit = list(lambda = 1e-06),
  ...
)
}
\arguments{
\item{x}{A numeric matrix representing the design matrix.  No missing valus
are allowed.  The coefficient estimates for constant columns will be
zero.  Thus, one should set the argument \code{intercept} to \code{TRUE}
to include an intercept term instead of adding an all-one column to
\code{x}.}

\item{treatment}{The assigned treatments represented by a character,
integer, numeric, or factor vector.}

\item{reward}{A numeric vector representing the rewards.  It is assumed that
a larger reward is more desirable.}

\item{propensity_score}{A numeric vector taking values between 0 and 1
representing the propensity score.}

\item{intercept}{A logical value indicating if an intercept should be
considered in the model.  The default value is \code{TRUE} and the
intercept is excluded from regularization.}

\item{weight}{A numeric vector for nonnegative observation weights. Equal
observation weights are used by default.}

\item{loss}{A character value specifying the loss function.  The available
options are \code{"logistic"} for the logistic deviance loss,
\code{"boost"} for the exponential loss approximating Boosting machines,
\code{"hinge-boost"} for hybrid of SVM and AdaBoost machine, and
\code{"lum"} for largin-margin unified machines (LUM).  See Liu, et
al. (2011) for details.}

\item{control}{A list of control parameters. See \code{abclass.control()}
for details.}

\item{nstages}{A positive integer specifying for the number of stages in the
ET-Lasso procedure.  By default, two rounds of tuning by random
permutations will be performed as suggested in Yang, et al. (2019).}

\item{refit}{A logical value indicating if a new classifier should be
trained using the selected predictors or a named list that will be
passed to \code{abclass.control()} to specify how the new classifier
should be trained.}

\item{...}{Other arguments passed to the control function, which calls the
\code{abclass.control()} internally.}
}
\description{
Tune the regularization parameter for MOML by the ET-Lasso method (Yang, et
al., 2019).
}
\references{
Yang, S., Wen, J., Zhan, X., & Kifer, D. (2019). ET-Lasso: A new efficient
tuning of lasso-type regularization for high-dimensional data. In
\emph{Proceedings of the 25th ACM SIGKDD International Conference on
Knowledge Discovery & Data Mining} (pp. 607--616).
}
