% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex.R
\name{vertex}
\alias{vertex}
\title{Simplex Vertices for The Angle-Based Classification}
\usage{
vertex(k)
}
\arguments{
\item{k}{Number of classes, a positive integer that is greater than one.}
}
\value{
A \code{(k-1)} by \code{k} matrix that consists of vertices in
    columns.
}
\description{
Simplex Vertices for The Angle-Based Classification
}
\references{
Lange, K., & Tong Wu, Tong (2008). An MM algorithm for multicategory vertex
discriminant analysis. Journal of Computational and Graphical Statistics,
17(3), 527--544.
}
