% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abclass-package.R
\docType{package}
\name{abclass-package}
\alias{abclass-package}
\title{Multi-Category Angle-Based Large-Margin Classifiers}
\description{
This package provides implementations of the multi-category angle-based
classifiers (Zhang & Liu, 2014) with the large-margin unified machines (Liu,
et al., 2011) for high-dimensional data.
}
\references{
Zhang, C., & Liu, Y. (2014). Multicategory Angle-Based Large-Margin
Classification. \emph{Biometrika}, 101(3), 625--640.

Liu, Y., Zhang, H. H., & Wu, Y. (2011). Hard or soft classification?
large-margin unified machines. \emph{Journal of the American Statistical
Association}, 106(493), 166--177.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wwenjie.org/abclass}
  \item \url{https://github.com/wenjie2wang/abclass}
  \item Report bugs at \url{https://github.com/wenjie2wang/abclass/issues}
}

}
\author{
\strong{Maintainer}: Wenjie Wang \email{wang@wwenjie.org} (\href{https://orcid.org/0000-0003-0363-3180}{ORCID})

Authors:
\itemize{
  \item Wenjie Wang \email{wang@wwenjie.org} (\href{https://orcid.org/0000-0003-0363-3180}{ORCID})
}

Other contributors:
\itemize{
  \item  Eli Lilly and Company [copyright holder]
}

}
\keyword{internal}
