% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abclass-package.R
\docType{package}
\name{abclass-package}
\alias{abclass-package}
\title{Multi-Category Angle-Based Large-Margin Classifiers}
\description{
This package provides implementations of the multi-category angle-based
classifiers (Zhang & Liu, 2014) with the large-margin unified machines (Liu,
et al., 2011) for high-dimensional data.
}
\references{
Zhang, C., & Liu, Y. (2014). Multicategory Angle-Based Large-Margin
Classification. \emph{Biometrika}, 101(3), 625--640.

Liu, Y., Zhang, H. H., & Wu, Y. (2011). Hard or soft classification?
large-margin unified machines. \emph{Journal of the American Statistical
Association}, 106(493), 166--177.
}
