% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abclass.R
\docType{package}
\name{abclass}
\alias{abclass}
\title{Multi-Category Angle-Based Large-Margin Classifiers}
\description{
The package malc provides implementations of multi-category angle-based
large-margin classifiers for high-dimensional data.
}
\references{
Zhang, C., & Liu, Y. (2014). Multicategory Angle-Based Large-Margin
Classification. \emph{Biometrika}, 101(3), 625--640.

Liu, Y., Zhang, H. H., & Wu, Y. (2011). Hard or soft classification?
large-margin unified machines. \emph{Journal of the American Statistical
Association}, 106(493), 166--177.
}
