% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.abclass}
\alias{coef.abclass}
\title{Coefficient Estimates of A Trained Angle-Based Classifier}
\usage{
\method{coef}{abclass}(object, selection = c("cv_min", "cv_1se", "all"), ...)
}
\arguments{
\item{object}{An object of class \code{abclass}.}

\item{selection}{A character value specifying how to select a particular set
of coefficient estimates from the entire solution path.  If the spcified
\code{abclass} object contains the cross-validation results, one may set
\code{selection} to \code{"cv_min"} (or \code{"cv_1se"}) for the
estimates giving the smallest cross-validation error (or the set of
estimates resulted from the largest \emph{lambda} within one standard
error of the smallest cross-validation error).  The entire solution path
will be returned in an array if \code{selection = "all"} or no
cross-validation results are available in the input \code{abclass}
object.}

\item{...}{Other arguments not used now.}
}
\value{
A vector representing the predictions or an array representing the
    entire solution path.
}
\description{
Extract coefficient estimates from an \code{abclass} object.
}
