% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.abclass}
\alias{predict.abclass}
\title{Prediction by A Trained Angle-Based Classifier}
\usage{
\method{predict}{abclass}(
  object,
  newx,
  type = c("class", "probability"),
  selection = c("cv_min", "cv_1se", "all"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{abclass}.}

\item{newx}{A numeric matrix representing the design matrix for predictions.}

\item{type}{A character value specifying the desired type of predictions.
The available options are \code{"class"} for predicted labels and
\code{"probability"} for class conditional probability estimates.}

\item{selection}{A character value specifying how to select a particular set
of coefficient estimates from the solution path for the predictions.
All the}

\item{...}{Other arguments not used now.}
}
\value{
A list containing the predictions.
}
\description{
Predict class labels or estimate conditional probabilities for the specified
new data.
}
