% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.supclass}
\alias{predict.supclass}
\title{Predictions}
\usage{
\method{predict}{supclass}(
  object,
  newx,
  type = c("class", "probability"),
  selection = c("cv_min", "cv_1se", "all"),
  ...
)
}
\description{
Predictions
}
